ignore size hints
Although I like size hints, they make no sense in a tiled and tabbed
window manager.  They only make sense when the size of a single window
depends only on a single client and a single client dictates the size of
a single window.  When the size of a window depends on the size of other
windows (as in a tiled situation), or when a set of clients must have
the same size (as in a tabbed situation), it makes no sense to constrain
the size of the window based on the size hints of a single client,
because the relation from clients to window size is no more one-to-one.